Subject Line: Fix memory leak in Game.cpp.
There was a memory leak caused by unallocated memory in Game.cpp.

First, `std::vector<Mesh*>*` meshObjects was changed
to no longer be a pointer. It was not necessary to have
that new/delete pairing as a class member. meshObjects is
now simply `std::vector<Mesh*>`.

Second, the Mesh* stored in meshObjects were not being
destroyed; std::vector::clear() does NOT destroy contents
of a vector when the contents are raw pointers.

Third, more updates can be made to this, namely, using shared_ptr
or unique_ptr in order to automatically manage the memory
allocated during the Mesh rendering process.

Changelist
- meshObjects changed to std::vector<Mesh*> (from std::vector<Mesh*>*)
- Memory leak fixed.

If an issue tracker is being used, place references to them at the bottom, like this:

[Resolves: #1]